column no should be an int
git-svn-id: http://svn.xp-dev.com/svn/wxextension@2699 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/extension/include/wx/extension/listview.h
@@ -169,7 +169,7 @@ public:
       wxListView::SetItemImage(item_number, iconid): false);};
 
   /// Sets the item text using item number and column number.
-  void SetItemText(long item_number, long col_number, const wxString& text);
+  void SetItemText(long item_number, int col_number, const wxString& text);
 
   /// Sorts on a column.
   /// If you specified use_images,
